+++ Backend/src/v1/__init__.py
@@ -1,6 +1,6 @@
 from .ping import router as ping_router
 from .data import router as data_router
-from .collections import router as collections_router
+from .dataset import router as collections_router
 
 
 __all__ = ["ping_router", "data_router", "collections_router"]
